- Issue #3 Removed occurences of patent expiration data from the data scraper file.
Updated description parsing to detect for "\n\n" empty descriptions and replace them with empty strings rather than "N/A". This will help with search functions later on.

Making the description an empty string doesn't make it show in the current data visualization. We can either add "N/A" in the visualization part, or later exclude "N/A" from the search feature.